📌(backend) pin Django to less than 3.0
Django 3.0 is out but not supported yet by DjangoCMS so we need to
make sure we stay on Django 2.x versions. This constraint was added
to richie v1.16 but we need to fix it here before we are able to
upgrade to v1.16 which requires some refactoring on instart's
templates and css.